Divide DI to other files
